Two New Features
Support passing a file as a response for testing.  The filename
is the value of an env var.  Mainly for unit testing and does not
support string vs object testing yet.

Support for new getter config value to specify the filename of a script
to extract the quote value when the response data is too complex for a
regex.

Signed-off-by: Chris. Webster <chris@webstech.net>