Add disable amdppm service
The amdppm power management software has a bug that appears to randomly
shut down the current processor. It causes the system to hang.  This
patch introduces a script to run at startup.

It will check the registry and if the service is started, will update
the registtry and reboot in a couple of minutes.  This is to catch when
the registry may have been updated by windows update.

Signed-off-by: Chris. Webster <chris@webstech.net>